readme ND4J

+++ demo07/src/main/java/com/codingapi/deeplearning/demo07/learn/DenseLayer.java
@@ -93,7 +93,7 @@ public class DenseLayer implements NeuralNetworkLayer {
      * 初始化权重参数
      */
     @Override
-    public void init(double lambda,double alpha,long seed) {
+    public void init(double lambda,double alpha,int seed) {
         this.lambda = lambda;
         this.alpha = alpha;
