+++ demo07/src/main/java/com/codingapi/deeplearning/demo07/learn/NeuralNetworkLayerBuilder.java
@@ -59,7 +59,7 @@ public class NeuralNetworkLayerBuilder {
     /**
      * 初始化所有层的权重 w,b
      */
-    public void init(double lamdba,double alpha,long seed) {
+    public void init(double lamdba,double alpha,int seed) {
         if(!noOutLay){
             throw new RuntimeException("没有输出层");
         }
