+++ demo08/src/main/java/com/codingapi/deeplearning/demo08/learn/core/NeuralNetwork.java
@@ -46,7 +46,7 @@ public class NeuralNetwork implements Serializable {
     private LossFunction lossFunction;
 
 
-    protected NeuralNetwork(double lambda, double alpha, int numEpochs,long seed,
+    protected NeuralNetwork(double lambda, double alpha, int numEpochs,int seed,
                          NeuralNetworkLayerBuilder layerBuilder,LossFunction lossFunction) {
         this.numEpochs = numEpochs;
         this.layerBuilder = layerBuilder;
