+++ demo09/src/main/java/com/codingapi/deeplearning/demo09/learn/layer/NeuralNetworkLayer.java
@@ -30,7 +30,7 @@ public interface NeuralNetworkLayer extends Serializable {
     /**
      * 初始化参数
      */
-    void init(double lamdba,double alpha,long seed);
+    void init(double lamdba,double alpha,int seed);
 
     /**
      * 权重值
