+++ demo10/src/main/java/com/codingapi/deeplearning/demo10/learn/core/NeuralNetworkBuilder.java
@@ -26,7 +26,7 @@ public class NeuralNetworkBuilder {
     private double lambda;
     private double alpha;
     private int numEpochs;
-    private long seed;
+    private int seed;
 
     private InputType inputType;
 
@@ -59,7 +59,7 @@ public class NeuralNetworkBuilder {
         return this;
     }
 
-    public NeuralNetworkBuilder seed(long seed){
+    public NeuralNetworkBuilder seed(int seed){
         this.seed = seed;
         return this;
     }
